+++ beam-core/Database/Beam/Backend/Types.hs
@@ -15,7 +15,6 @@ module Database.Beam.Backend.Types
 
 import           Control.Monad.Free.Church
 import           Control.Monad.Identity
-import qualified Data.Aeson as Json
 import           Data.Tagged
 import           Data.Vector.Sized (Vector)
 import qualified Data.Vector.Sized as Vector