+++ docs/beam-templates/employee1out-agg.hs
@@ -37,7 +37,7 @@ data ShoppingCartDb f = ShoppingCartDb
                       { _shoppingCartUsers :: f (TableEntity UserT) }
                         deriving Generic
 
-instance Database ShoppingCartDb
+instance Database be ShoppingCartDb
 
 shoppingCartDb :: DatabaseSettings Sqlite ShoppingCartDb
 shoppingCartDb = defaultDbSettings