+++ docs/beam-templates/chinookdml.hs
@@ -36,6 +36,8 @@ main =
      stmts <- newIORef id
 
      let onStmt s = modifyIORef stmts (. (s:))
+
+         record :: BEAM_BACKEND_MONAD a -> IO a
          record a = withDatabaseDebug (onStmt . (++ ";")) chinook a
 
      handle (\BeamDone -> pure ()) $