+++ beam-sqlite/ChangeLog.md
@@ -1,3 +1,7 @@
+# 0.3.1.0
+
+Add `runBeamSqlite` and `runBeamSqliteDebug` functions
+
 # 0.3.0.0
 
 * Re-introduce backend parameter to `Database` class